requirements.txt
Django==3.2
mysqlclient
django_pandas
matplotlib
sklearn
